Audit fixes & e2e tests (#23)
* disable tests

* SB10 - block duplicate token in dual migration

* SB09 - fix SettlementCache params order & use object format

* SB14 - use forceApprove()

* SB13 - support chain IDs larger than type(uint32).max

* SB12 - remove dangling approval & not over-send ether

* SB11 & SB08 - reentrancy guard

* SB04 - strengthen settlement cache key

* SB03 - remove amounts swap (assume token routes are in order)

* SB18 - refund (and all transfer) using WETH instead of ETH

* SB17 - remove magic numbers

* SB01 - early return for zero address

* SB20 - immutable Uniswap addresses

* SB19 - remove settlement data check, as it's implicitly done via migration hash

* SB16 - cache value for gas efficiency

* SB04 addendum - revert on amount too low or unexpected token

* enumerate test cases

get first test working again

UniswapV3AcrossMigrator tests

added routeMin check

* libraries tests

chain settlers & protocol fees

cleanup

migrator

Settler tests

Settler tests - touchup

* UniswapV3AcrossSettler single token path tests

UniswapV3AcrossSettler Dual Token tests

* migrators tests

settlers tests

* license

* Remaining e2e tests

* make UniswapV4Library.mintPosition public for size optimization

* emit Refund event every time

* add refund event check in e2e tests

* MigrationHash -> MigrationId
Remove catching specific reverts from try catch

* updated tests for previous commit

* draft audit report

* updated readme and default .env.example

* final audit report

* add final owner to deploy script

---------

+++ src/base/AcrossMigrator.sol
@@ -1,4 +1,4 @@
-// SPDX-License-Identifier: UNLICENSED
+// SPDX-License-Identifier: MIT
 pragma solidity ^0.8.24;
 
 import {IERC20} from "@openzeppelin/token/ERC20/IERC20.sol";
@@ -36,7 +36,7 @@ abstract contract AcrossMigrator is IAcrossMigrator, Migrator {
     /// @param data The data to bridge
     function _bridge(
         address sender,
-        uint32 chainId,
+        uint256 chainId,
         address settler,
         address token,
         uint256 amount,
