+++ test/mocks/MockUniswapV4Migrator.sol
@@ -14,7 +14,7 @@ contract MockUniswapV4Migrator is UniswapV4Migrator {
 
     function _bridge(
         address sender,
-        uint32 chainId,
+        uint256 chainId,
         address settler,
         address token,
         uint256 amount,
